Fix webpack manifest paths
thanks to SUI using `context: 'client/'` in webpack,
all the paths reported by the plugin are off by a level,
so ui-classic has `./node_modules/` in the manifest,
but ui-service has `../node_modules`, leading to a failure in package lookup.

Moving the common code to a separate file (that's still 1:1 with ui-classic),
and applying the path patch to modules.

Fixes ManageIQ/manageiq-ui-classic#7390
