+++ config/webpack.prod.js
@@ -4,56 +4,9 @@ const HtmlWebpackPlugin = require('html-webpack-plugin')
 const OptimizeCssAssetsWebpackPlugin = require('optimize-css-assets-webpack-plugin')
 
 const { StatsWriterPlugin } = require('webpack-stats-plugin');
-const fs = require('fs');
-const sortBy = require('lodash/sortBy');
 
-const config = require('./webpack.dev.js')
-
-const uniqueWebpackModules = (data) => {
-  // recursively flatten nested modules
-  const transformModule  = m  => m.modules ? transformModules(m.modules) : m.name.split("!").pop();
-  const transformModules = ms => ms.flatMap(m => transformModule(m));
-
-  var modules = transformModules(data.modules).sort();
-  modules = [...new Set(modules)]; // uniq
-
-  return modules;
-}
-
-const packagesFromModules = (modules) => {
-  // find the package.json file for each module
-  modules = modules.filter(m => m.includes("node_modules"));
-  var packagePaths = modules.map(m => {
-    var match = m.match(/(.*node_modules\/)([^/]+)(\/[^/]+)?/);
-    var path = [
-      `${match[1]}${match[2]}/package.json`,
-      `${match[1]}${match[2]}${match[3]}/package.json`,
-    ].find(p => fs.existsSync(p));
-
-    if (path == null) {
-      console.warn(`[webpack-manifest] WARN: Unable to find a package.json for ${m}`);
-    }
-
-    return path;
-  })
-  packagePaths = packagePaths.filter(p => p != null);
-  packagePaths = [...new Set(packagePaths)]; // uniq
-
-  // extract relevant package data from the package.json
-  var packages = packagePaths.map(p => {
-    var content = fs.readFileSync(p);
-    var pkg = JSON.parse(content);
-    return {
-      name: pkg.name,
-      license: pkg.license,
-      version: pkg.version,
-      location: p
-    }
-  });
-  packages = sortBy(packages, ['name', 'version']);
-
-  return packages;
-}
+const config = require('./webpack.dev.js');
+const manifest = require('./manifest.js');
 
 // Source maps suitable for production use
 config.devtool = 'cheap-module-source-map'
@@ -87,7 +40,7 @@ config.plugins.push(
     fields: ["modules"],
 
     transform(data) {
-      var modules = uniqueWebpackModules(data);
+      const modules = manifest.uniqueWebpackModules(data).map((path) => path.replace(/^\.\.\//, './'));
       return JSON.stringify(modules, null, 2);
     }
   }),
@@ -98,8 +51,8 @@ config.plugins.push(
     fields: ["modules"],
 
     transform(data) {
-      var modules = uniqueWebpackModules(data);
-      var packages = packagesFromModules(modules);
+      const modules = manifest.uniqueWebpackModules(data).map((path) => path.replace(/^\.\.\//, './'));
+      const packages = manifest.packagesFromModules(modules);
       return JSON.stringify(packages, null, 2);
     }
   })
